+++ src/ui/control/geolocate_control.test.ts
@@ -468,14 +468,31 @@ describe('GeolocateControl with no options', () => {
         map.jumpTo({
             center: [10, 20]
         });
+
+        // test with bugger radius
         let zoomendPromise = map.once('zoomend');
-        map.zoomTo(10, {duration: 0});
-        await zoomendPromise;
-        expect(geolocate._circleElement.style.width).toBe('20px'); // 700m = 20px at zoom 10
-        zoomendPromise = map.once('zoomend');
         map.zoomTo(12, {duration: 0});
         await zoomendPromise;
-        expect(geolocate._circleElement.style.width).toBe('79px'); // 700m = 79px at zoom 12
+        expect(geolocate._circleElement.style.width).toBe('79px');
+        console.log(geolocate._circleElement.style.width);
+        zoomendPromise = map.once('zoomend');
+        map.zoomTo(10, {duration: 0});
+        await zoomendPromise;
+        expect(geolocate._circleElement.style.width).toBe('20px');
+        console.log(geolocate._circleElement.style.width);
+        zoomendPromise = map.once('zoomend');
+
+        // test with smaller radius
+        geolocation.send({latitude: 10, longitude: 20, accuracy: 20});
+        map.zoomTo(20, {duration: 0});
+        await zoomendPromise;
+        expect(geolocate._circleElement.style.width).toBe('19982px');
+        console.log(geolocate._circleElement.style.width);
+        zoomendPromise = map.once('zoomend');
+        map.zoomTo(18, {duration: 0});
+        await zoomendPromise;
+        expect(geolocate._circleElement.style.width).toBe('4996px');
+        console.log(geolocate._circleElement.style.width);
     });
 
     test('shown even if trackUserLocation = false', async () => {