+++ src/ui/map/requestRenderFrame.test.ts
@@ -6,11 +6,27 @@ beforeEach(() => {
 
 describe('requestRenderFrame', () => {
 
-    test('Map#_requestRenderFrame schedules a new render frame if necessary', () => {
+    test('Map#_requestRenderFrame schedules a new render frame if necessary', (done) => {
         const map = createMap();
         const spy = jest.spyOn(map, 'triggerRepaint');
         map._requestRenderFrame(() => {});
-        expect(spy).toHaveBeenCalledTimes(1);
+        expect(spy).toHaveBeenCalledTimes(0);
+
+        // wait for style to be loaded
+        map.once('data', () => {
+            spy.mockReset();
+            map._requestRenderFrame(() => {});
+            expect(spy).toHaveBeenCalledTimes(1);
+            map.remove();
+            done();
+        });
+    });
+
+    test('Map#_requestRenderFrame should not schedule a render frame before style load', () => {
+        const map = createMap();
+        const spy = jest.spyOn(map, 'triggerRepaint');
+        map._requestRenderFrame(() => {});
+        expect(spy).toHaveBeenCalledTimes(0);
         map.remove();
     });